Copy file mounts rather than move them
